fix: rename embedded search folder, move all related code to the folder
