+++ app/component/embedded/EmbeddedSearch.js
