+++ app/component/embedded/EmbeddedSearchContainer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { matchShape } from 'found';
-import LazilyLoad, { importLazy } from './LazilyLoad';
+import LazilyLoad, { importLazy } from '../LazilyLoad';
 
 const modules = {
   EmbeddedSearch: () => importLazy(import('./EmbeddedSearch')),
