+++ app/component/embedded/embedded-search-generator.scss
