+++ app/component/embedded/embedded-search.scss
