+++ app/component/embedded/index.js
